Computer World - CRUD Implementation & More
Implemented CRUD for products table which takes in Product code, name, and price. A very basic instance of a foreign key has been created. The manufacturers id  from the manufacturers table is being used in this table. It currently takes no data, I just needed to test if it worked.

MORE REFACTORING NEEDED.