Computer World - Database Changes & Foreign Keys
I have altered the database design quite a bit. I have 3 tables:
Product.
Manufacturer,
Order

Each of them link to one another without looping. Product is used to store the products in the store. Each product is linked to manufacturer. Order communicates with the Product table and receives the product id which we will use to receive the product prices which will in turn give us the total price of the order.

This has yet to be implemented. The foreign keys do link up without issues so far. I'll now focus on getting data input into the system.

More refactoring will be done.