Computer World - Refactored names & folders
Removed some files from the project and placed that code into their respective files. May change this again later.

Renamed all of the files and folders to make them more readable and similar to the MDM project that was presented to us.

Added the 'try' test folder which will make it easy to test large parts of my code.

Foreign keys have been created and are linked correctly. Data not displaying correctly in Postman.