Computer World - Foreign Keys Functioning
Implemented the pre-loading function from GORM.

When we use  GET in postman, it displays the data of that record and shows the data that is linked to the foreign key (manufacturer data) to refactor the code to be more efficient and readable.

Need to set-up the orders table.

I'll begin testing implementation soon.