Computer World - Validation & Repositories
Implemented a good chunk of the feedback given from Cameron and Matt.
Introduced a repositories folder which contains access to each table. After the data has been validated and checked for errors. It is then sent to its corresponding repository file where the business section of the code takes place e.g. Create, Get, Update, and Delete.

Create is currently the only functioning part of the code, as I wanted to make sure the new code was functioning before adding more.

Introduced an error response function based on feedback which is used to return an error

The data created in the requests files are now being used in the controllers

Will now continue to fix the rest of the problems .