Computer World - Fully Implemented CRUD
The controllers and repositories are now communicating with each other with no known problems so far. CRUD is fully implemented using requests, responses, validation, and repositories.

Need to implement more tests or a way to retrieve all records in a table when get with no id. I wIll look at the MDM-api for more details.