Computer World - Implemented More Tests
Added almost double the amount of tests. All of them run and return the correct status. I will implement a system which prints of the testing information soon, this is so we can see that the tests aren't broken or being cheesed.

Need to apply more validation methods for the update functions in the controller.