Computer World - Bug Fixes & Changes
The front-end was broken due to the coding changes in the API. I have since managed to fix these problems and they no longer cause any problems, or at least any problems that I know of.

All tests currently pass, either in bulk or individually. The front-end is working as intended. The error handling and validation on the API is working.