Computer World - Project Overhaul Part 1
I have completely re-written a majority of the files. The api is currently non-functioning due to a variety of errors.

Endpoints for the user handler
Added migrations for all tabls
Rewritten all models to now use UIDs
Renamed quite a few files and folders to match that of the mdm api.
Added factories for all models.
A lot more.